+++ Sources/CorePayments/Networking/GetFundingEligibilityVariables.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct GetFundingEligibilityVariables: Encodable {
+
+    let clientId: String
+    let intent: String
+    let currency: String
+    let enableFunding: [String]
+}