+++ Sources/CorePayments/UIApplication+URLOpener.swift
@@ -4,6 +4,7 @@ import UIKit
 public protocol URLOpener {
     func open(_ url: URL, completionHandler completion: ((Bool) -> Void)?)
     func isPayPalAppInstalled() -> Bool
+    func isVenmoAppInstalled() -> Bool
 }
 
 extension UIApplication: URLOpener {
@@ -14,6 +15,12 @@ extension UIApplication: URLOpener {
         }
         return canOpenURL(payPalURL)
     }
+
+    // swiftlint:disable force_unwrapping
+    public func isVenmoAppInstalled() -> Bool {
+        canOpenURL(URL(string: "com.venmo.touch.v2://")!)
+    }
+    // swiftlint:enable force_unwrapping
 
     public func open(_ url: URL, completionHandler completion: ((Bool) -> Void)?) {
         UIApplication.shared.open(url, options: [:], completionHandler: completion)