+++ UnitTests/TestShared/MockURLOpener.swift
@@ -4,6 +4,7 @@ import Foundation
 class MockURLOpener: URLOpener {
 
     var mockIsPayPalAppInstalled = false
+    var mockIsVenmoAppInstalled = false
     var mockOpenURLSuccess = true
     var lastOpenedURL: URL?
 
@@ -11,6 +12,10 @@ class MockURLOpener: URLOpener {
 
     func isPayPalAppInstalled() -> Bool {
         return mockIsPayPalAppInstalled
+    }
+
+    func isVenmoAppInstalled() -> Bool {
+        return mockIsVenmoAppInstalled
     }
 
     func open(_ url: URL, completionHandler completion: ((Bool) -> Void)?) {