+++ UnitTests/VenmoPaymentsTests/Environment+Venmo_Tests.swift
@@ -1,0 +1,28 @@
+import XCTest
+@testable import CorePayments
+@testable import VenmoPayments
+
+class Environment_Venmo_Tests: XCTestCase {
+
+    func testVenmoEnvironment_sandbox_returnsCorrectCheckoutURL() {
+        XCTAssertEqual(
+            Environment.sandbox.venmoCheckoutBaseURL.absoluteString,
+            "https://www.sandbox.paypal.com/smart/checkout/venmo"
+        )
+    }
+
+    func testVenmoEnvironment_live_returnsCorrectCheckoutURL() {
+        XCTAssertEqual(
+            Environment.live.venmoCheckoutBaseURL.absoluteString,
+            "https://www.paypal.com/smart/checkout/venmo"
+        )
+    }
+
+    func testVenmoEnvironment_sandbox_returnsCorrectEnvironmentString() {
+        XCTAssertEqual(Environment.sandbox.venmoEnvironmentString, "sandbox")
+    }
+
+    func testVenmoEnvironment_live_returnsCorrectEnvironmentString() {
+        XCTAssertEqual(Environment.live.venmoEnvironmentString, "production")
+    }
+}